+++ internal/provider/resource_check_ssh.go
@@ -57,7 +57,7 @@ type CheckSSHResourceModel struct {
 	IsPaused               types.Bool   `tfsdk:"is_paused"`
 	Interval               types.Int64  `tfsdk:"interval"`
 	NumRetries             types.Int64  `tfsdk:"num_retries"`
-	UseIpVersion           types.String `tfsdk:"use_ip_version"`
+	UseIPVersion           types.String `tfsdk:"use_ip_version"`
 	Notes                  types.String `tfsdk:"notes"`
 	IncludeInGlobalMetrics types.Bool   `tfsdk:"include_in_global_metrics"`
 	SLA                    types.Object `tfsdk:"sla"`
@@ -101,7 +101,7 @@ func (a CheckSSHResourceModelAdapter) ToAPIArgument(model CheckSSHResourceModel)
 		IsPaused:               upapi.BoolPtr(model.IsPaused.ValueBool()),
 		Interval:               model.Interval.ValueInt64(),
 		NumRetries:             model.NumRetries.ValueInt64(),
-		UseIpVersion:           model.UseIpVersion.ValueString(),
+		UseIPVersion:           stringOptionalAPIValue(model.UseIPVersion),
 		Notes:                  model.Notes.ValueString(),
 		IncludeInGlobalMetrics: upapi.BoolPtr(model.IncludeInGlobalMetrics.ValueBool()),
 	}
@@ -132,7 +132,7 @@ func (a CheckSSHResourceModelAdapter) FromAPIResult(api upapi.Check) (*CheckSSHR
 		IsPaused:               types.BoolValue(api.IsPaused),
 		Interval:               types.Int64Value(api.Interval),
 		NumRetries:             types.Int64Value(api.NumRetries),
-		UseIpVersion:           types.StringValue(api.UseIPVersion),
+		UseIPVersion:           types.StringValue(api.UseIPVersion),
 		Notes:                  types.StringValue(api.Notes),
 		IncludeInGlobalMetrics: types.BoolValue(api.IncludeInGlobalMetrics),
 		SLA: a.SLAAttributeValue(SLAAttribute{